app/requirements.txt
torch
ultralytics
supervision
opencv-python-headless
streamlit
lap
norfair
torchvision
opencv-python
pandas
Pillow